fix(eve): split dynamic tool reference analysis
